suggest Holochain for project spotlight
I've never done this before, so I don't know what the conventions and etiquette are about this section :) Suggesting because Holochain recently reached its first beta, which may be a noteworthy event, not sure. If this PR is rejected, I'd love feedback on how to approach this differently. Thanks!

(Disclaimer: I'm a technical writer for the Holochain project.)